tools/server-check: support distros with postgres ident authentication
In such distros, we have to run once as root, and once as the postgres
user, which can access the db, not any other file.

+++ tools/server-check.sh
@@ -10,24 +10,53 @@ get_ini ()
 }
 
 DEVEL_MODE=
-if [ "$1" == "-d" ] ; then
-    DEVEL_MODE=y
-    echo "Devel mode!"
-fi
+SUCMD=
+DB_HOST=
+DB_PORT=
+DB_USER=openerp
+DB_PASSWORD=
+IN_SU=
+
+PG_ROOT=postgres
+
+while [ -n "$1" ] ; do
+    case "$1" in
+    -d)
+	DEVEL_MODE=y
+	echo "Devel mode!"
+	;;
+    -s)
+	SUCMD="su $PG_ROOT - "
+	;;
+    -h)
+	DB_HOST=$2
+	shift 1
+	;;
+    -p)
+	DB_PORT=$2
+	shift 1
+	;;
+    -U)
+	DB_USER=$2
+	shift 1
+	;;
+    -W)
+	DB_PASSWORD=$2
+	shift 1
+	;;
+    --in-su)
+	IN_SU=y
+	;;
+    esac
+    shift 1
+done
 
 if [ ! -f "/var/run/openerp-server-check" ] && [ -z "$DEVEL_MODE" ] ; then
     # only run this if the magic file is there
     exit 0
 fi
 
-DB_HOST=
-DB_PORT=
-DB_USER=openerp
-DB_PASSWORD=
-
-PG_ROOT=postgres
-
-if [ -n "$DEVEL_MODE" ] && [ -f ~/openerp-server.conf ] ; then
+if [ -n "$DEVEL_MODE" ] && [ -r ~/openerp-server.conf ] ; then
     echo "Parsing" ~/openerp-server.conf
     get_ini ~/openerp-server.conf
 elif [ -f "/etc/openerp-server.conf" ] ; then
@@ -51,10 +80,24 @@ if [ -n "$DB_PORT" ] ; then
     DB_CONNS+=" --port $DB_PORT"
 fi
 
+if [ -n "$SUCMD"  ] ; then
+    CMD="$0 --in-su"
+    if [ -n "$DB_HOST" ] ; then 
+	CMD="$CMD -h $DB_HOST"
+    fi
+    if [ -n "$DB_PORT" ] ; then
+	CMD="$CMD -p $DB_PORT"
+    fi
+    if [ -n "$DB_PASSWORD" ] ; then
+	CMD="$CMD -W $DB_PASSWORD"
+    fi
+    su $PG_ROOT -c "$CMD"
+    exit "$?"
+fi
 
 if ! (psql -qt -U $PG_ROOT $DB_CONNS -c "SELECT usename FROM pg_user WHERE usename = '$DB_USER';" | \
 	grep $DB_USER > /dev/null) ; then
-	if ! createuser -U $PG_ROOT $DB_CONNS -S -d -R -l $DB_USER < /dev/null ; then
+	if ! $SUCMD createuser -U $PG_ROOT $DB_CONNS -S -d -R -l $DB_USER < /dev/null ; then
 		echo "Failed to create user $DB_USER"
 		exit 1
 	fi
@@ -63,7 +106,7 @@ else
 fi
 
 echo "OK"
-if [ -z "$DEVEL_MODE" ] ; then
+if [ -z "$DEVEL_MODE" ] && [ -z "$IN_SU" ] ; then
     rm -f "/var/run/openerp-server-check"
 fi
